feat: isolate advanced to its own page
