+++ apps/multisig/src/layouts/NewTransaction/NewTransactionHeader.tsx
@@ -1,3 +1,3 @@
 export const NewTransactionHeader: React.FC<React.PropsWithChildren> = ({ children }) => (
-  <h2 css={({ color }) => ({ color: color.offWhite, marginTop: 4, textAlign: 'center' })}>{children}</h2>
+  <h2 css={({ color }) => ({ color: color.offWhite, marginTop: 4 })}>{children}</h2>
 )
